Fixed COCOA sniffer for MacOS X; moved internal modules in a specific package.
