+++ selfspy/modules/period.py
